requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.3.2
Flask==3.0.0
frozenlist==1.4.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector-python==8.2.0
protobuf==4.21.12
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.4
